+++ crates/astria-sequencer-relayer/tests/blackbox/helpers/mock_sequencer_server.rs
@@ -20,6 +20,8 @@ use astria_core::{
         sequencerblock::v1::{
             GetUpgradesInfoRequest,
             GetUpgradesInfoResponse,
+            GetValidatorNameRequest,
+            GetValidatorNameResponse,
         },
     },
     primitive::v1::RollupId,
@@ -144,6 +146,13 @@ impl SequencerService for SequencerServiceImpl {
         self: Arc<Self>,
         _request: Request<GetUpgradesInfoRequest>,
     ) -> Result<Response<GetUpgradesInfoResponse>, Status> {
+        unimplemented!()
+    }
+
+    async fn get_validator_name(
+        self: Arc<Self>,
+        _request: Request<GetValidatorNameRequest>,
+    ) -> Result<Response<GetValidatorNameResponse>, Status> {
         unimplemented!()
     }
 }