+++ crates/astria-sequencer/src/authority/mod.rs
@@ -50,7 +50,7 @@ impl ValidatorSet {
         self.0.get(address.address_bytes())
     }
 
-    pub(super) fn push_update(&mut self, update: ValidatorUpdate) {
+    pub(super) fn insert(&mut self, update: ValidatorUpdate) {
         self.0
             .insert(*update.verification_key.address_bytes(), update);
     }