+++ system-tests/sequencer_upgrade_test.py
@@ -13,6 +13,7 @@ For details on running the test, see the README.md file in `/system-tests`.
 import argparse
 import concurrent
 import aspen_upgrade_checks
+import time
 from concurrent.futures import FIRST_EXCEPTION
 from helpers.astria_cli import Cli
 from helpers.evm_controller import EvmController
@@ -84,6 +85,14 @@ done, _ = concurrent.futures.wait(futures, return_when=FIRST_EXCEPTION)
 for completed_future in done:
     completed_future.result()
 
+# Instantiate CLI
+cli = Cli(upgrade_image_tag)
+
+# Convert node addresses to astria-prefixed bech32m addresses.
+for node in nodes:
+    address = node.address
+    node.bech32m_address = cli.address(node.name, address)
+
 # Note block 1 and the current app version before attempting the upgrade.
 for node in nodes:
     node.wait_until_chain_at_height(1, 60)
@@ -108,6 +117,15 @@ for node in nodes[1:]:
     if genesis_app_version != node.get_app_version_at_genesis():
         raise SystemExit(f"node0 and {node.name} report different values for genesis app version")
 
+# Run pre-upgrade validator updates to check that the new action still executes correctly.
+print("running pre-upgrade validator updates")
+for node in nodes:
+    node.power += 1
+    cli.validator_update(node.name, node.pub_key, node.power)
+
+# Give time for validator updates to land
+time.sleep(2)
+
 # Run pre-upgrade checks specific to this upgrade.
 print(f"running pre-upgrade checks specific to {upgrade_name}")
 if upgrade_name == "aspen":
@@ -121,7 +139,6 @@ print("testing bridge in")
 evm_balance = evm.get_balance(EVM_DESTINATION_ADDRESS)
 if evm_balance != 0:
     raise SystemExit(f"starting evm balance not 0: balance {evm_balance}")
-cli = Cli(upgrade_image_tag)
 cli.init_bridge_account(sequencer_name="node1")
 cli.bridge_lock(sequencer_name="node2")
 expected_evm_balance = 10000000000000000000
@@ -189,6 +206,7 @@ new_node.deploy_sequencer(
     upgrade_name=upgrade_name,
     upgrade_activation_height=upgrade_activation_height
 )
+new_node.bech32m_address = cli.address(new_node.name, new_node.address)
 
 # Wait for the new node to catch up and go through the upgrade too.
 new_node.wait_until_chain_at_height(upgrade_activation_height + 2, 60)
@@ -234,6 +252,11 @@ for node in nodes[1:]:
             f"node0 and {node.name} report different values for scheduled upgrade changes"
         )
 print("upgrade change infos reported correctly")
+
+# Submit validator updates with names for all validators
+print("running post-upgrade validator updates")
+for node in nodes:
+    cli.validator_update(node.name, node.pub_key, node.power)
 
 # Run post-upgrade checks specific to this upgrade.
 print(f"running post-upgrade checks specific to {upgrade_name}")